move code to lib/initialization.dart
