+++ lib/web_ui/lib/ui.dart
@@ -34,5 +34,4 @@ part 'pointer.dart';
 part 'semantics.dart';
 part 'text.dart';
 part 'tile_mode.dart';
-part 'util.dart';
 part 'window.dart';
